refactor(channel): extract mouse hit-testing + geometry into channel_mouse.go
Stacks on the sidebar extraction. Pulls mouseAction type, the two-tier
hit-tester (mouseActionAt -> mainPanelMouseAction), main-panel geometry
projector (mainPanelGeometry), and the recovery-prompt click handler
(applyRecoveryPrompt) into a dedicated kin file. mouseAction is the
shape every dispatch path produces; co-locating its type with the
producer keeps the protocol legible.

  - cmd/wuphf/channel.go: -214 LOC (4463 -> 4249)
  - cmd/wuphf/channel_mouse.go: +242 LOC (new)

Methods moved:
  - mouseAction (type)
  - mouseActionAt (top-level dispatch over layout)
  - mainPanelMouseAction (within-panel dispatch over rows / cards / popups)
  - mainPanelGeometry (header / message / composer / popup row counts)
  - applyRecoveryPrompt (handler for mouseAction Kind="prompt")

No behavior change. All existing tests pass.

Cumulative channel.go reduction across the stack:
  4807 -> 4687 (pickers)   -120
       -> 4463 (sidebar)    -224
       -> 4249 (mouse)      -214
  Net: -558 LOC; channel.go shrunk by 12% so far in Phase 2.
